+++ server/handlers/notify.js
@@ -1,21 +1,37 @@
-import Promise from "bluebird";
-import { smartNotifierHandler } from "hull/lib/utils";
-import updateUser from "../lib/update-user";
+const _ = require("lodash");
+const Promise = require("bluebird");
+const { smartNotifierHandler } = require("hull/lib/utils");
+
+const updateUser = require("../lib/update-user");
 
 const notify = smartNotifierHandler({
   handlers: {
     "user:update": (ctx, messages = []) => {
       const { smartNotifierResponse } = ctx;
+      const concurrency = parseInt(
+        _.get(ctx, "ship.private_settings.concurrency", 10),
+        10
+      );
+      if (_.isNaN(concurrency)) {
+        concurrency = 10;
+      }
       // Get 10 users every 100ms at most.
       smartNotifierResponse.setFlowControl({
         type: "next",
         size: 10,
         in: 100
       });
-      const upd = updateUser.bind(undefined, ctx);
-      return Promise.all(messages.map(upd));
+      return Promise.map(
+        messages,
+        message => {
+          return updateUser(ctx, message);
+        },
+        {
+          concurrency
+        }
+      );
     }
   }
 });
 
-export default notify;
+module.exports = notify;