+++ server/handlers/status.js
@@ -1,6 +1,6 @@
-import _ from "lodash";
+const _ = require("lodash");
 
-export default function statusCheck(req, res) {
+function statusCheck(req, res) {
   const { ship, client } = req.hull;
   const { private_settings = {} } = ship;
   const {
@@ -39,3 +39,5 @@ export default function statusCheck(req, res) {
   res.json({ messages, status });
   return client.put(`${req.hull.ship.id}/status`, { status, messages });
 }
+
+module.exports = statusCheck;