+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.2.9
+- add concurrency support
+- upgrade hull-node
+- cleanup babeljs and flow configuration, rewrite import/export syntax
+- replace request with superagent
+
 ## 0.2.8
 
 - re-add smart-notifier